block: Pass filename to bdrv_get_full_backing_filename()
RH-Author: Jeffrey Cody <jcody@redhat.com>
Message-id: <38864625eb0b9eab7854303d40fdc286bb743621.1393613946.git.jcody@redhat.com>
Patchwork-id: 57947
O-Subject: [RHEL6.6 qemu-kvm PATCH] block: Pass filename to bdrv_get_full_backing_filename()
Bugzilla: 1067576
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>

This is an interim fix for a regression introduced by:

    commit 387ea7cb8345bdcc900ffe145cbcdcf65da91257
    qemu-img: make "info" backing file output correct and easier to use

In that commit, bdrv_get_full_backing_filename() was introduced,
which replaced calling path_combine() manually.  The difference
is that rather than using the filename string as passed to
bdrv_open(), it uses the filename string attached to the BDS.

Both the backing_file and filename strings in the BDS are limited
to 1024 characters.  The backing_file string built up in bdrv_open(),
however, has a limit of PATH_MAX, which is 4096 under Linux.

This difference comes into play when using an image chain that has
a medium-to-large number of images, all of which have relative-pathed
backing file references to the parent directory.

For instance, consider the following backing chain:

tstA
├── base.qcow2
├── sn1.qcow2   (backing file ../tstA/base.qcow2)
├── sn2.qcow2   (backing file ../tstA/sn1.qcow2)
└── sn3.qcow2   (backing file ../tstA/sn2.qcow2)

The backing filename string is built up with the relative paths intact,
like so:

    base.qcow2:  ../tstA../tstA../tstA/base.qcow2

The backing filename is then passed into the bdrv_open() call to open
the backing file.

When using lv volume names, the snapshot and pathname ends up longer,
and after ~23 snapshots we have hit or exceeded the 1024 byte limit
for the BDS filename.

If the filename is run through the BDS, it is truncated at 1024
characters.  Then the open for the backing file fails, as we are
building the filename from a truncated string.

By passing the filename directly to bdrv_get_full_backing_filename(),
we can support a built-up string size of 4096 bytes.

This is not a long-term solution, because a character limit of 4096
bytes will be hit with additional images.  The proper long-term
solution should happen upstream first, and consist of:

1) dynamically allocated filename strings in the BDS
2) flattening redundant relative pathname strings

But as this is a customer-reported regression issue that they are running
into, a temporary z-stream solution will provide similar behavior to
RHEL6.4.

Since this is not the final solution, and the fix really is relevant
just to undo a regression, the fix is downstream only.  It will be
replaced by the final upstream fix, once complete.

BZ 1067576
Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=7124797
---
 block.c    | 7 ++++---
 block.h    | 2 +-
 qemu-img.c | 2 +-
 3 files changed, 6 insertions(+), 5 deletions(-)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>