block: gluster - add reopen support.
RH-Author: Jeffrey Cody <jcody@redhat.com>
Message-id: <2f1e357ca27ef45d16e62fbe44b0ce9cef2602fe.1394129674.git.jcody@redhat.com>
Patchwork-id: 58041
O-Subject: [RHEL7 qemu-kvm PATCH 2/2] block: gluster - add reopen support.
Bugzilla: 1031526
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

Gluster does parse open flags in its .bdrv_open() implementation,
and the .bdrv_reopen_* implementations need to do the same.

A new gluster connection to the image file to be created is established
in the .bdrv_reopen_prepare(), and the image file opened with the new
flags.

If this is successful, then the old image file is closed, and the
old connection torn down. The relevant structure pointers in the gluster
state structure are updated to the new connection.

If it is not successful, then the new file handle and connection is
abandoned (if it exists), while the old connection is not modified at
all.

With reopen supported, block-commit (and offline commit) is now also
supported for image files whose base image uses the native gluster
protocol driver.

Signed-off-by: Jeff Cody <jcody@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
(cherry picked from commit adccfbcd6020e928db93b2b4faf0dbd05ffbe016)

Conflicts:
	block/gluster.c

RHEL7 Note:  Conflict was in contextual information, as RHEL7 does not
             have gluster zerofill support.
---
 block/gluster.c | 102 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 1 file changed, 102 insertions(+)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>