spec: Add README.rhel6-gpxe-source
RH-Author: Eduardo Habkost <ehabkost@redhat.com>
Message-id: <1394217468-2864-2-git-send-email-ehabkost@redhat.com>
Patchwork-id: 58049
O-Subject: [RHEL7 qemu-kvm PATCH 1/2] spec: Add README.rhel6-gpxe-source
Bugzilla: 1073774
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Amos Kong <akong@redhat.com>

Bugzilla: 1073774
Brew scratch build: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=7157997

We need to provide users to the source code for the RHEL6 gPXE images we
ship, to follow the GPL. This file will document how to obtain the
source code. The file is also important for us, so we know where those
images came from.

Note: A previous patch sent by Orit (but never included) documented the
files as coming from gpxe-roms-qemu-0.9.7-6.10.el6, but that was
incorrect. I found out that the files are from
gpxe-roms-qemu-0.9.7-6.9.el6.

To be on the safe side, I provided two different methods to obtain the
source code: a simple ftp.redhat.com URL, and a pointer to the
corresponding chapter on the RHEL Installation Guide that describes
where and how to obtain the source code ISOs.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
---
 redhat/Makefile.common          | 2 +-
 redhat/README.rhel6-gpxe-source | 9 +++++++++
 redhat/qemu-kvm.spec.template   | 4 +++-
 3 files changed, 13 insertions(+), 2 deletions(-)
 create mode 100644 redhat/README.rhel6-gpxe-source

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>