net: Update netdev peer on link change
RH-Author: Vlad Yasevich <vyasevic@redhat.com>
Message-id: <1396367023-2433-1-git-send-email-vyasevic@redhat.com>
Patchwork-id: 58322
O-Subject: [RHEL6.6 qemu-kvm PATCH] net: Update netdev peer on link change
Bugzilla: 965396
RH-Acked-by: Xiao Wang <jasowang@redhat.com>
RH-Acked-by: Amos Kong <akong@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1000375
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=7284587
Upstream: 02d38fcb2caa4454cf4ed728d5908c3cc9ba47be

When a link change occurs on a backend (like tap), we currently do
not propage such change to the nic.  As a result, when someone turns
off a link on a tap device, for instance, then a guest doesn't see
that change and continues to try to send traffic or run DHCP even
though the lower-layer is disconnected.  This is OK when the network
is set up as a HUB since the the guest may be connected to other HUB
ports too, but when it's set up as a netdev, it makes thinkgs worse.

The patch addresses this by setting the peers link down only when the
peer is not a HUBPORT device.  With this patch, in the following config
  -netdev tap,id=net0 -device e1000,mac=XXXXX,netdev=net0
when net0 link is turned off, the guest e1000 shows lower-layer link
down. This allows guests to boot much faster in such configurations.
With windows guest, it also allows the network to recover properly
since windows will not configure the link-local IPv4 address, and
when the link is turned on, the proper address address is configured.

Signed-off-by: Vlad Yasevich <vyasevic@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>

Conflicts:
    net.c  [context and different file]

Signed-off-by: Vlad Yasevich <vyasevic@redhat.com>
---
 net.c | 24 +++++++++++++++---------
 1 file changed, 15 insertions(+), 9 deletions(-)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>