qcow2: Don't rely on free_cluster_index in alloc_refcount_block() (CVE-2014-0147)
RH-Author: Stefan Hajnoczi <stefanha@redhat.com>
Message-id: <1396448702-20414-1-git-send-email-stefanha@redhat.com>
Patchwork-id: n/a
O-Subject: [RHEL-6.6/6.5.z qemu-kvm PATCH v3] qcow2: Don't rely
           on free_cluster_index in alloc_refcount_block() (CVE-2014-0147)
Bugzilla: 1079338
RH-Acked-by: Max Reitz <mreitz@redhat.com>
RH-Acked-by: Jeff Cody <jcody@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1079338
Upstream status: b106ad9185f35fc4ad669555ad0e79e276083bd7

free_cluster_index is only correct if update_refcount() was called from
an allocation function, and even there it's brittle because it's used to
protect unfinished allocations which still have a refcount of 0 - if it
moves in the wrong place, the unfinished allocation can be corrupted.

So not using it any more seems to be a good idea. Instead, use the
first requested cluster to do the calculations. Return -EAGAIN if
unfinished allocations could become invalid and let the caller restart
its search for some free clusters.

The context of creating a snapsnot is one situation where
update_refcount() is called outside of a cluster allocation. For this
case, the change fixes a buffer overflow if a cluster is referenced in
an L2 table that cannot be represented by an existing refcount block.
(new_table[refcount_table_index] was out of bounds)

Signed-off-by: Kevin Wolf <kwolf@redhat.com>

Conflicts:
	block/qcow2-refcount.c
	block/qcow2.c
	tests/qemu-iotests/044.out
	tests/qemu-iotests/080
	tests/qemu-iotests/080.out

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>