PCI: Convert pci_info() to QObject
RH-Author: Marcel Apfelbaum <marcel.a@redhat.com>
Message-id: <1401952158-29608-2-git-send-email-marcel.a@redhat.com>
Patchwork-id: 59147
O-Subject: [RHEL-6.6 qemu-kvm PATCH v5 1/7] PCI: Convert pci_info() to QObject
Bugzilla: 813748
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Luiz Capitulino <lcapitulino@redhat.com>

The returned QObject is a QList of all buses. Each bus is
represented by a QDict, which has a key with a QList of all
PCI devices attached to it. Each device is represented by
a QDict.

As has happended to other complex conversions, it's hard to
split this commit as part of it are new functions which are
called by each other.

IMPORTANT: support for printing PCI bridge attached devices
is NOT part of this commit, it's going to be added by the
next commit, as it's untested.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
(cherry picked from commit 163c8a59f6f9c370b74a81856194757e3773b358)

Conflicts:
	hw/pci.c
	hw/pci.h

RHEL6 Notes:
 - both conflicts were because of include lines mismatch

Signed-off-by: Marcel Apfelbaum <marcel.a@redhat.com>
---
 hw/pci.c  | 370 +++++++++++++++++++++++++++++++++++++++++++++++++-------------
 hw/pci.h  |   3 +-
 monitor.c |   3 +-
 3 files changed, 300 insertions(+), 76 deletions(-)

Signed-off-by: jen <jen@redhat.com>