kvmclock: Ensure time in migration never goes backward
RH-Author: Marcelo Tosatti <mtosatti@redhat.com>
Message-id: <20140626150716.560273759@amt.cnet>
Patchwork-id: 59388
O-Subject: [RHEL-7.1 qemu-kvm PATCH 1/2] kvmclock: Ensure time in migration never goes backward
Bugzilla: 1098602
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>

commit a096b3a6732f846ec57dc28b47ee9435aa0609bf upstream
Author: Alexander Graf <agraf@suse.de>
Date:   Fri May 16 17:15:21 2014 +0200

When we migrate we ask the kernel about its current belief on what the guest
time would be. However, I've seen cases where the kvmclock guest structure
indicates a time more recent than the kvm returned time.

To make sure we never go backwards, calculate what the guest would have seen
as time at the point of migration and use that value instead of the kernel
returned one when it's more recent.  This bases the view of the kvmclock
after migration on the same foundation in host as well as guest.

Signed-off-by: Alexander Graf <agraf@suse.de>
Cc: qemu-stable@nongnu.org
Reviewed-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
BZ: 1098602

---
 hw/i386/kvm/clock.c |   49 +++++++++++++++++++++++++++++++++++++++++++++++++
 1 file changed, 49 insertions(+)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>