qmp: improve debuggability of BLOCK_IO_ERROR event
Message-id: <1404916312-27953-4-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 59730
O-Subject: [RHEV7.1 qemu-kvm-rhev PATCH 3/3] qmp: improve debuggability of BLOCK_IO_ERROR event
Bugzilla: 1116772
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Eric Blake <eblake@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

This commit forward ports the following RHEL-7.0 commit to RHEL-7.1:

  commit bfea65d6f229fd65c78ae4daaddf0d0711d0aedd
  Author: Laszlo Ersek <lersek@redhat.com>
  Date:   Thu Nov 21 16:27:19 2013 +0100

  improve debuggability of BLOCK_IO_ERROR / BLOCK_JOB_ERROR (RHEL 6->7 fwd)

I had to redo the work because now events use the QAPI, but it
was straightforward.

There's one significant difference though: this commit does not
extend the BLOCK_JOB_ERROR event as did the RHEL-7.0 commit. The
reason is that the original requirement was to have this only on
the BLOCK_IO_ERROR event. Let's not spread it unless we're asked to.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
---
 block.c                 | 16 ++++++++++++++--
 docs/qmp/qmp-events.txt | 10 +++++++++-
 qapi/block-core.json    | 18 +++++++++++++++++-
 3 files changed, 40 insertions(+), 4 deletions(-)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>