block: vhdx - remove BAT file offset bit shifting
RH-Author: Jeffrey Cody <jcody@redhat.com>
Message-id: <6c55c6a5a5bb64eb0ca033447f49157eb948c80d.1405033137.git.jcody@redhat.com>
Patchwork-id: 59759
O-Subject: [RHEL6-6 qemu-kvm PATCH v2 14/24] block: vhdx - remove BAT file offset bit shifting
Bugzilla: 1063559
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

Bit shifting can be fun, but in this case it was unnecessary.  The
upper 44 bits of the 64-bit BAT entry is specifies the File Offset,
so we shifted the bits to get access to the value.

However, per the spec the value is in MB.  So we dutifully shifted back
to the left by 20 bits, to convert to a true uint64_t file offset.

This replaces those steps with just a bit mask, to get rid of the lower
20 bits instead.

Signed-off-by: Jeff Cody <jcody@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
(cherry picked from commit 0b7da092b40734538631c3ad461c1753a87535fc)
---
 block/vhdx.c | 6 ++----
 block/vhdx.h | 1 -
 2 files changed, 2 insertions(+), 5 deletions(-)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>