block: vhdx - improve error message, and .bdrv_check implementation
RH-Author: Jeffrey Cody <jcody@redhat.com>
Message-id: <fd950f635c638edfa7be271d6315ff943bc52fa5.1405033137.git.jcody@redhat.com>
Patchwork-id: 59767
O-Subject: [RHEL6-6 qemu-kvm PATCH v2 22/24] block: vhdx - improve error message, and .bdrv_check implementation
Bugzilla: 1063559
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

If there is a dirty log file to be replayed in a VHDX image, it is
replayed in .vhdx_open().  However, if the file is opened read-only,
then a somewhat cryptic error message results.

This adds a more helpful error message for the user.  If an image file
contains a log to be replayed, and is opened read-only, the user is
instructed to run 'qemu-img check -r all' on the image file.

Running qemu-img check -r all will cause the image file to be opened
r/w, which will replay the log file.  If a log file replay is detected,
this is flagged, and bdrv_check will increase the corruptions_fixed
count for the image.

[Fixed typo in error message that was pointed out by Eric Blake
<eblake@redhat.com>.
--Stefan]

Signed-off-by: Jeff Cody <jcody@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
(cherry picked from commit 7e30e6a6746b417c7e0dbc9af009560fbb63f336)

RHEL6 Notes:

    Differences from upstream due to lack of Error argument to
    vhdx_open().  Instead of setting errp, vhdx_parse_log() will
    print the error message via error_report, if there is a log
    to replay and the image was opened r/o.

Signed-off-by: Jeff Cody <jcody@redhat.com>
---
 block/vhdx-log.c |  9 +++++++++
 block/vhdx.c     | 22 ++++++++++++++++++++--
 block/vhdx.h     |  2 ++
 3 files changed, 31 insertions(+), 2 deletions(-)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>