simpletrace: add simpletrace.py --no-header option
RH-Author: Stefan Hajnoczi <stefanha@redhat.com>
Message-id: <1405099110-10975-6-git-send-email-stefanha@redhat.com>
Patchwork-id: 59862
O-Subject: [RHEL6.6 qemu-kvm PATCH 5/7] simpletrace: add simpletrace.py --no-header option
Bugzilla: 905761
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

It can be useful to read simpletrace files that have no header.  For
example, a ring buffer may not have a header record but can still be
processed if the user is sure the file format version is compatible.

  $ scripts/simpletrace.py --no-header trace-events trace-file

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
---
 scripts/simpletrace.py | 24 +++++++++++++++++-------
 1 file changed, 17 insertions(+), 7 deletions(-)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>