migration: dump vmstate info as a json file for static analysis
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <ce3ffaa556efc2422222ba6eaa925768cab0e3ff.1405075344.git.amit.shah@redhat.com>
Patchwork-id: 59801
O-Subject: [RHEL6.6 qemu-kvm PATCH 01/18] migration: dump vmstate info as a json file for static analysis
Bugzilla: 1118718
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Dr. David Alan Gilbert (git) <dgilbert@redhat.com>

This commit adds a new command, '-dump-vmstate', that takes a filename
as an argument.  When executed, QEMU will dump the vmstate information
for the machine type it's invoked with to the file, and quit.

The JSON-format output can then be used to compare the vmstate info for
different QEMU versions, specifically to test whether live migration
would break due to changes in the vmstate data.

A Python script that compares the output of such JSON dumps is included
in the following commit.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Reviewed-by: Juan Quintela <quintela@redhat.com>
Signed-off-by: Juan Quintela <quintela@redhat.com>
(cherry picked from commit abfd9ce341ec66eb2e63756b9da43f77c054788e)
Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: jen <jen@redhat.com>

Conflicts:
	hw/hw.h
	qemu-options.hx
	vl.c

Changes for RHEL6:
* VMS_MUST_EXIST doesn't exist in RHEL6, so dropped flags check in
  savevm.c.
* current_machine not accessible in savevm.c due to hw/boards.h being
  platform-dependent.  Have to pass it in via vl.c instead.
* Walking device tree is slightly different in savevm.c.
* qemu-options.hx: command definitions take one argument less
  (QEMU_ARCH_ALL dropped).
---
 hw/hw.h         |   2 +
 qemu-options.hx |  13 ++++++
 savevm.c        | 126 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 vl.c            |  14 +++++++
 4 files changed, 155 insertions(+)