Include OHCI device for ppc64
Message-id: <1405917919-7547-2-git-send-email-dgibson@redhat.com>
Patchwork-id: 59980
O-Subject: [PATCH 1/3] Include OHCI device for ppc64
Bugzilla:
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>

With the current RHEL7 qemu configuration, UHCI, EHCI and XHCI devices are
included on all platforms, but not OHCI.  But on real hardware pp64
machines (and in fact most non-x86 platforms), OHCI is the standard USB1
controller, and UHCI is almost unheard of.

Arguably we should replace UHCI with OHCI on all non-x86 platforms.  But
for now just enable OHCI for ppc64, where we definitely need it.

Signed-off-by: David Gibson <dgibson@redhat.com>
---
 default-configs/ppc64-softmmu.mak | 1 +
 1 file changed, 1 insertion(+)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>