Use separately packaged SLOF on ppc64
For ppc64 targets, qemu needs the SLOF (Slim-Line Open Firmware) firmware
image.  At the moment we're including the pre-built binary from the qemu
tree in our qemu package for ppc64.

But including an un-sourced binary isn't nice, so following Fedora's lead
we want to have SLOF built in its own package.  This changes the qemu
packaging not to include the SLOF binary, but instead to require the
separate SLOF package on ppc64.

Signed-off-by: David Gibson <dgibson@redhat.com>