scsi-disk: improve the lba-out-of-range tests for read/write/verify
RH-Author: Markus Armbruster <armbru@redhat.com>
Message-id: <1406020565-25364-3-git-send-email-armbru@redhat.com>
Patchwork-id: 59992
O-Subject: [PATCH 6.6 qemu-kvm v2 02/11] scsi-disk: improve the lba-out-of-range tests for read/write/verify
Bugzilla: 1064643
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: Ronnie Sahlberg <ronniesahlberg@gmail.com>

Improve the tests for the LBA to cover more cases.

For the 16 byte opcodes, the lba is a uint64, so we need to check is to
make sure that we do not wrap.  For example if an opcode would specify
the LBA:0xffffffffffffffff and LEN:2 then lba+len would wrap to 1.

Also verify that ALL requested blocks are available, not just the first one.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit ba6095cd6b497783e3621606b6b2320781aa3f52)
Signed-off-by: Markus Armbruster <armbru@redhat.com>
---
 hw/scsi-disk.c | 6 ++++--
 1 file changed, 4 insertions(+), 2 deletions(-)

Signed-off-by: jen <jen@redhat.com>