block: gluster - add reopen support
RH-Author: Jeffrey Cody <jcody@redhat.com>
Message-id: <b6eec65bbdb9f1e33e57c24c817ba4d845499977.1406658609.git.jcody@redhat.com>
Patchwork-id: 60308
O-Subject: [RHEL-6.6 qemu-kvm PATCH 2/2] block: gluster - add reopen support
Bugzilla: 1118543
RH-Acked-by: Max Reitz <mreitz@redhat.com>
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

Gluster does parse open flags in its .bdrv_open() implementation,
and the .bdrv_reopen_* implementations need to do the same.

A new gluster connection to the image file to be created is established
in the .bdrv_reopen_prepare(), and the image file opened with the new
flags.

If this is successful, then the old image file is closed, and the
old connection torn down. The relevant structure pointers in the gluster
state structure are updated to the new connection.

If it is not successful, then the new file handle and connection is
abandoned (if it exists), while the old connection is not modified at
all.

With reopen supported, block-commit (and offline commit) is now also
supported for image files whose base image uses the native gluster
protocol driver.

Signed-off-by: Jeff Cody <jcody@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
(cherry picked from commit adccfbcd6020e928db93b2b4faf0dbd05ffbe016)
Signed-off-by: jen <jen@redhat.com>

Conflicts:
	block/gluster.c

[ RHEL-6 Notes: conflict due to context after the new code, due to
                gluster in rhel6 not having zerofill support ]

Signed-off-by: Jeff Cody <jcody@redhat.com>
---
 block/gluster.c | 102 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 1 file changed, 102 insertions(+)