vmschk-check-rhel6.py: quickly compare RHEL6 jsons based on criteria
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <affc5a565f732d75a76dd5bd2958617de7e74a59.1406890029.git.amit.shah@redhat.com>
Patchwork-id: 60394
O-Subject: [RHEL6.6 qemu-kvm PATCH 2/2] vmschk-check-rhel6.py: quickly compare RHEL6 jsons based on criteria
Bugzilla: 1125792
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Dr. David Alan Gilbert (git) <dgilbert@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

This script compares the RHEL6 jsons added in the following commit for a
variety of criteria: various machine types and various source-destination
combinations, while providing the option to enable or disable backward
migration.

The default is to compare all the RHEL6 jsons found in
tests/vmstate-static-checker-data.  Backward migration checks (src >
dest) are enabled with the -b switch.

The tests can be restricted to a particular source json (-s), or a
particular destination json (-d), or to a particular machine type (-m).

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 scripts/vmschk-check-rhel6.py | 139 ++++++++++++++++++++++++++++++++++++++++++
 1 file changed, 139 insertions(+)
 create mode 100755 scripts/vmschk-check-rhel6.py

Signed-off-by: jen <jen@redhat.com>