RHEL only: Disable tests that don't work with RHEL build on ppc
Message-id: <1410147705-12993-3-git-send-email-dgibson@redhat.com>
Patchwork-id: 60892
O-Subject: [PATCH qemu-kvm-rhev RHEV7.1 2/5] RHEL only: Disable tests that don't work with RHEL build on ppc
Bugzilla: 1113998
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

Some of the ppc testsuite cases rely on the PReP and various Macintosh and
embedded machine types, which we don't support in the RHEL build of
qemu-kvm-rhev.  We disable those tests so that the rest of make check can
work properly.

Signed-off-by: David Gibson <dgibson@redhat.com>

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>