RHEL onlyy: Disable unused ppc machine types
Message-id: <1410147705-12993-4-git-send-email-dgibson@redhat.com>
Patchwork-id: 60893
O-Subject: [PATCH qemu-kvm-rhev RHEV7.1 3/5] RHEL onlyy: Disable unused ppc machine types
Bugzilla: 1113998
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

The default qemu configuration for ppc64 includes a number of machine
types we don't care about for RHEL or RHEV.  This include PReP,
old-style ppc based Macintosh machines, and several ppc4xx and BookE
based embedded boards.

Some can be disabled with the config file, others require makefile hacks.

With this patch, only the "pseries" machine type remains for ppc which
is the only one we care about for RHEV.

Note that this only disables the machine types, it doesn't avoid
compilation of individual devices which are specific to those
machines.  That's a job for another day.

Signed-off-by: David Gibson <dgibson@redhat.com>
---
 default-configs/ppc64-softmmu.mak | 6 ------
 hw/ppc/Makefile.objs              | 2 +-
 2 files changed, 1 insertion(+), 7 deletions(-)

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>